% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cranalmanac.R
\name{cran_rxiv_pkg_children}
\alias{cran_rxiv_pkg_children}
\title{Get all CRAN archived packages, version numbers, and when they were last modified}
\usage{
cran_rxiv_pkg_children(cran_rxiv_pkg_parent_tbl, n = 100)
}
\arguments{
\item{n}{Number of packages to retrieve, the default is 100, if you leave this as NULL it will try (and possibly fail) to retrieve all packages}
}
\description{
Get all CRAN archived packages, version numbers, and when they were last modified
}
