% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cranalmanac.R
\name{cran_rxiv_pkg_name_link}
\alias{cran_rxiv_pkg_name_link}
\title{Get the list of all the names of all the links}
\usage{
cran_rxiv_pkg_name_link(cran_rxiv_pkg_parent_tbl)
}
\arguments{
\item{cran_rxiv_pkg_parent_tbl}{the table of parent package names}
}
\value{
list of urls of all archived CRAN packages
}
\description{
Get the list of all the names of all the links
}
\examples{

\dontrun{
cran_rxiv_pkg_parent_tbl <- cran_rxiv_pkgs()
cran_rxiv_pkg_name_link(cran_rxiv_pkg_parent_tbl)
}

}
