% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranalmanac.R
\name{cran_rxiv_pkg_parent}
\alias{cran_rxiv_pkg_parent}
\title{Get CRAN archived package names and when they were last modified}
\usage{
cran_rxiv_pkg_parent()
}
\value{
a data.frame of the pkg name, and when it was last modified
}
\description{
Get CRAN archived package names and when they were last modified
}
\examples{

\dontrun{
cran_rxiv_pkg_parent_tbl <- cran_rxiv_pkgs()
}
}
