% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cranalmanac.R
\name{tidy_scraped_cran_rxiv}
\alias{tidy_scraped_cran_rxiv}
\title{Clean up the output of cran_rxiv_pkg_children}
\usage{
tidy_scraped_cran_rxiv(cran_rxiv_pkg_children)
}
\arguments{
\item{cran_rxiv_pkg_children}{}
}
\value{
data.frame with columns
}
\description{
This function is used inside
}
